ttl: split metrics into separate file
Release note: None
